rag/scripts/requirements.txt
sentence-transformers
faiss-cpu
python-docx
pdfminer.six
numpy
tqdm
openai
